Move `PlotTrace`, `timeAxis` classes to new file
